+++ src/api/mod.rs
@@ -1,6 +1,8 @@
 pub mod eastmoney;
 pub mod eastmoney_error;
+pub mod eastmoney_helpers;
 pub mod eastmoney_ranking;
+pub mod eastmoney_types;
 pub mod f10_apidata;
 pub mod f10_jbgk;
 pub mod fund_holdings;